move helpers into conftest
