+++ .flake8
@@ -1,3 +1,3 @@
 [flake8]
-ignore = W191,W503
+ignore = W191,W503,E203
 max-line-length = 240
